# spec/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/pico-units/{picoUnitId}/control/setpoints:
    put:
      description: >-
        Modify the temperature and/or humidity targets for a Pico Unit. If the control loop is enabled, these changes
        will affect the devices. If success, a new reading will be generated afterwards.
      operationId: PicoUnitIdControlController_setpoints_v1
      parameters:
        - name: picoUnitId
          required: true
          in: path
          description: Pico unit ID
          schema:
            type: integer
            minimum: 1
            example: 42
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeSetPointsDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetpointsDto'
        '404':
          description: PicoUnit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '406':
          description: Database-related error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '410':
          description: This action can only be implemented with enabled Pico Units
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '417':
          description: A service returned an error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 422
                  error:
                    type: string
                    example: Unprocessable Entity
                  message:
                    type: string
                    example:
                      - propertyA has wrong value "...", [constrains]
                      - propertyB has wrong value "...", [constrains]
                      - '...'
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '424':
          description: Axios produced an error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Generic unknown error produced by an unidentified exception
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  error:
                    type: string
                    example: Internal Server Error
                  message:
                    type: string
                    example: Internal server error
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '502':
          description: A request was made but no response was received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '503':
          description: The event loop is lagging and no more requests can be attended right now
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 503
                  error:
                    type: string
                    example: Service Unavailable
                  message:
                    type: string
                    example: Server is too busy
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
      summary: Change temperature and/or humidity targets
      tags:
        - proxy
        - control
  /v1/pico-units/{picoUnitId}/control/setup:
    put:
      description: >-
        Use this endpoint to modify any of the PIN connections in the Pico Unit. Using it will make the Unit change its
        PIN-devices mapping. Use this endpoint with care.
      operationId: PicoUnitIdControlController_setup_v1
      parameters:
        - name: picoUnitId
          required: true
          in: path
          description: Pico unit ID
          schema:
            type: integer
            minimum: 1
            example: 42
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeSetupDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevicesDto'
        '404':
          description: PicoUnit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '406':
          description: Database-related error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '410':
          description: This action can only be implemented with enabled Pico Units
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '417':
          description: A service returned an error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 422
                  error:
                    type: string
                    example: Unprocessable Entity
                  message:
                    type: string
                    example:
                      - propertyA has wrong value "...", [constrains]
                      - propertyB has wrong value "...", [constrains]
                      - '...'
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '424':
          description: Axios produced an error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Generic unknown error produced by an unidentified exception
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  error:
                    type: string
                    example: Internal Server Error
                  message:
                    type: string
                    example: Internal server error
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '502':
          description: A request was made but no response was received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '503':
          description: The event loop is lagging and no more requests can be attended right now
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 503
                  error:
                    type: string
                    example: Service Unavailable
                  message:
                    type: string
                    example: Server is too busy
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
      summary: Change connections setup
      tags:
        - proxy
        - control
  /v1/pico-units/{picoUnitId}/control/outputs:
    put:
      description: >-
        Change the status of the Pico Unit's devices. This endpoint will only have an effect if the control loop is
        deactivated.
      operationId: PicoUnitIdControlController_outputs_v1
      parameters:
        - name: picoUnitId
          required: true
          in: path
          description: Pico unit ID
          schema:
            type: integer
            minimum: 1
            example: 42
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeOutputsDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutputsDto'
        '404':
          description: PicoUnit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '406':
          description: Database-related error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '409':
          description: This endpoint cannot be used as long as the Pico Unit's control loop is activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '410':
          description: This action can only be implemented with enabled Pico Units
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '417':
          description: A service returned an error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 422
                  error:
                    type: string
                    example: Unprocessable Entity
                  message:
                    type: string
                    example:
                      - propertyA has wrong value "...", [constrains]
                      - propertyB has wrong value "...", [constrains]
                      - '...'
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '424':
          description: Axios produced an error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Generic unknown error produced by an unidentified exception
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  error:
                    type: string
                    example: Internal Server Error
                  message:
                    type: string
                    example: Internal server error
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '502':
          description: A request was made but no response was received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '503':
          description: The event loop is lagging and no more requests can be attended right now
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 503
                  error:
                    type: string
                    example: Service Unavailable
                  message:
                    type: string
                    example: Server is too busy
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
      summary: Turn devices on and/or off
      tags:
        - proxy
        - control
  /v1/pico-units/{picoUnitId}/control/loop:
    put:
      description: >-
        Turns the Pico Unit's control loop (the logic that keeps track of the targets and turns devices on and off
        automatically to meet them) on and off. This action will override any manual settings that the outputs might be
        using and will use the temperature and humidity set points.
      operationId: PicoUnitIdControlController_loop_v1
      parameters:
        - name: picoUnitId
          required: true
          in: path
          description: Pico unit ID
          schema:
            type: integer
            minimum: 1
            example: 42
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControlLoopDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControlLoopDto'
        '404':
          description: PicoUnit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '406':
          description: Database-related error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '410':
          description: This action can only be implemented with enabled Pico Units
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '417':
          description: A service returned an error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 422
                  error:
                    type: string
                    example: Unprocessable Entity
                  message:
                    type: string
                    example:
                      - propertyA has wrong value "...", [constrains]
                      - propertyB has wrong value "...", [constrains]
                      - '...'
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '424':
          description: Axios produced an error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Generic unknown error produced by an unidentified exception
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  error:
                    type: string
                    example: Internal Server Error
                  message:
                    type: string
                    example: Internal server error
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '502':
          description: A request was made but no response was received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '503':
          description: The event loop is lagging and no more requests can be attended right now
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 503
                  error:
                    type: string
                    example: Service Unavailable
                  message:
                    type: string
                    example: Server is too busy
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
      summary: Turn control loop on or off
      tags:
        - proxy
        - control
info:
  title: mushpi-server
  description: A backend for the Mushroom Pi project
  version: 0.1.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    secret:
      type: http
      scheme: bearer
      in: header
      name: authorization
  schemas:
...
    ErrorDto:
      type: object
      properties: {}
...
    ChangeSetPointsDto:
      type: object
      properties:
        temperature:
          type: number
          description: The temperature in Celsius degrees that we want the Pico Unit to maintain
          example: 27
          minimum: 0
          maximum: 50
        humidity:
          type: number
          description: The percentage of humidity that we want the Pico Unit to maintain
          example: 60
          minimum: 20
          maximum: 90
    SetpointsDto:
      type: object
      properties: {}
    ChangeSetupDto:
      type: object
      properties:
        active_high:
          type: boolean
          description: >-
            A flag that should be true if the relay's circuits are connected to the NO (Normally Open) port, and false
            if they are connected through NC (Normally Closed).
    DevicesDto:
      type: object
      properties: {}
    ChangeOutputsDto:
      type: object
      properties:
        humidifier:
          type: boolean
          description: A flag that should be true we want the humidifier to be ON
        fan:
          type: boolean
          description: A flag that should be true we want the fan to be ON
        heater:
          type: boolean
          description: A flag that should be true we want the heater to be ON
    OutputsDto:
      type: object
      properties: {}
    ControlLoopDto:
      type: object
      properties:
        enabled:
          type: boolean
          description: >-
            A flag that should be true if we want the Pico Unit to automatically turn on and off devices to meet the
            temperature and humidity goals.
      required:
        - enabled